Actualizacion entidad paciente

+++ UI/frmProducto.cs
@@ -219,7 +219,7 @@ namespace UI
             txtCantidad.Value = productoSelected.cantidad;
             //obtiene el valor que el usuario selecciono en el combo
             cboFamilia.SelectedItem = (FamiliaProducto)productoSelected.familia;
-            cboCategoria.SelectedValue = (int)productoSelected.CategoriaId;
+            cboCategoria.SelectedValue = productoSelected.CategoriaId;
         }
 
         private void btnCancelar_Click(object sender, EventArgs e)
